Split mobile files
