+++ zcncore/transaction_mobile.go
@@ -1,0 +1,507 @@
+//go:build mobile
+// +build mobile
+
+package zcncore
+
+import (
+	"encoding/json"
+	"fmt"
+	"strconv"
+	"time"
+
+	"github.com/0chain/errors"
+	"github.com/0chain/gosdk/core/transaction"
+)
+
+type TransactionCommon interface {
+	// ExecuteSmartContract implements wrapper for smart contract function
+	ExecuteSmartContract(address, methodName string, input string, val string) error
+
+	// Send implements sending token to a given clientid
+	Send(toClientID string, val string, desc string) error
+	// SetTransactionFee implements method to set the transaction fee
+	SetTransactionFee(txnFee string) error
+
+	VestingAdd(ar *VestingAddRequest, value string) error
+
+	MinerSCLock(minerID string, lock string) error
+
+	FinalizeAllocation(allocID string, fee string) error
+	CancelAllocation(allocID string, fee string) error
+	CreateAllocation(car *CreateAllocationRequest, lock, fee string) error //
+	CreateReadPool(fee string) error
+	ReadPoolLock(allocID string, blobberID string, duration int64, lock, fee string) error
+	ReadPoolUnlock(poolID string, fee string) error
+	StakePoolLock(blobberID string, lock, fee string) error
+	StakePoolUnlock(blobberID string, poolID string, fee string) error
+	UpdateBlobberSettings(blobber *Blobber, fee string) error
+	UpdateAllocation(allocID string, sizeDiff int64, expirationDiff int64, lock, fee string) error
+	WritePoolLock(allocID string, blobberID string, duration int64, lock, fee string) error
+	WritePoolUnlock(poolID string, fee string) error
+}
+
+func parseCoinStr(vs string) (uint64, error) {
+	v, err := strconv.ParseUint(vs, 10, 64)
+	if err != nil {
+		return nil, fmt.Errorf("invalid token value: %v", vs)
+	}
+
+	if v/uint64(TOKEN_UNIT) == 0 {
+		return nil, fmt.Errorf("token value must be multiple value of 1e10")
+	}
+	return v, nil
+}
+
+// NewTransaction allocation new generic transaction object for any operation
+func NewTransaction(cb TransactionCallback, txnFee string, nonce int64) (TransactionScheme, error) {
+	err = CheckConfig()
+	if err != nil {
+		return nil, err
+	}
+	if _config.isSplitWallet {
+		if _config.authUrl == "" {
+			return nil, errors.New("", "auth url not set")
+		}
+		Logger.Info("New transaction interface with auth")
+		return newTransactionWithAuth(cb, fee, nonce)
+	}
+	Logger.Info("New transaction interface")
+	t, err := newTransaction(cb, fee, nonce)
+	return t, err
+}
+
+func (t *Transaction) ExecuteSmartContract(address, methodName string, input string, val string) error {
+	v, err := parseCoinStr(val)
+	if err != nil {
+		return err
+	}
+
+	err = t.createSmartContractTxn(address, methodName, json.RawMessage(input), v)
+	if err != nil {
+		return err
+	}
+	go func() {
+		t.setNonceAndSubmit()
+	}()
+	return nil
+}
+
+func (t *Transaction) SetTransactionFee(txnFee string) error {
+	fee, err := parseCoinStr(txnFee)
+	if err != nil {
+		return err
+	}
+
+	if t.txnStatus != StatusUnknown {
+		return errors.New("", "transaction already exists. cannot set transaction fee.")
+	}
+	t.txn.TransactionFee = fee
+	return nil
+}
+
+func (t *Transaction) Send(toClientID string, val string, desc string) error {
+	v, err := parseCoinStr(val)
+	if err != nil {
+		return err
+	}
+
+	txnData, err := json.Marshal(SendTxnData{Note: desc})
+	if err != nil {
+		return errors.New("", "Could not serialize description to transaction_data")
+	}
+	go func() {
+		t.txn.TransactionType = transaction.TxnTypeSend
+		t.txn.ToClientID = toClientID
+		t.txn.Value = v
+		t.txn.TransactionData = string(txnData)
+		t.setNonceAndSubmit()
+	}()
+	return nil
+}
+
+func (t *Transaction) SendWithSignatureHash(toClientID string, val string, desc string, sig string, CreationDate int64, hash string) error {
+	v, err := parseCoinStr(val)
+	if err != nil {
+		return err
+	}
+
+	txnData, err := json.Marshal(SendTxnData{Note: desc})
+	if err != nil {
+		return errors.New("", "Could not serialize description to transaction_data")
+	}
+	go func() {
+		t.txn.TransactionType = transaction.TxnTypeSend
+		t.txn.ToClientID = toClientID
+		t.txn.Value = v
+		t.txn.Hash = hash
+		t.txn.TransactionData = string(txnData)
+		t.txn.Signature = sig
+		t.txn.CreationDate = CreationDate
+		t.setNonceAndSubmit()
+	}()
+	return nil
+}
+
+func (t *Transaction) VestingAdd(ar *VestingAddRequest, value string) (
+	err error) {
+
+	v, err := parseCoinStr(value)
+	if err != nil {
+		return err
+	}
+
+	err = t.createSmartContractTxn(VestingSmartContractAddress,
+		transaction.VESTING_ADD, ar, v)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+func (t *Transaction) MinerSCLock(nodeID string, lock string) (err error) {
+	v, err := parseCoinStr(lock)
+	if err != nil {
+		return err
+	}
+
+	var mscl MinerSCLock
+	mscl.ID = nodeID
+
+	err = t.createSmartContractTxn(MinerSmartContractAddress,
+		transaction.MINERSC_LOCK, &mscl, v)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// FinalizeAllocation transaction.
+func (t *Transaction) FinalizeAllocation(allocID string, fee string) (
+	err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type finiRequest struct {
+		AllocationID string `json:"allocation_id"`
+	}
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_FINALIZE_ALLOCATION, &finiRequest{
+			AllocationID: allocID,
+		}, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// CancelAllocation transaction.
+func (t *Transaction) CancelAllocation(allocID string, fee string) (
+	err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type cancelRequest struct {
+		AllocationID string `json:"allocation_id"`
+	}
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_CANCEL_ALLOCATION, &cancelRequest{
+			AllocationID: allocID,
+		}, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// CreateAllocation transaction.
+func (t *Transaction) CreateAllocation(car *CreateAllocationRequest,
+	lock, fee string) (err error) {
+	lv, err := parseCoinStr(lock)
+	if err != nil {
+		return err
+	}
+
+	fv, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_CREATE_ALLOCATION, car, lv)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(fv)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// CreateReadPool for current user.
+func (t *Transaction) CreateReadPool(fee string) (err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_CREATE_READ_POOL, nil, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// ReadPoolLock locks tokens for current user and given allocation, using given
+// duration. If blobberID is not empty, then tokens will be locked for given
+// allocation->blobber only.
+func (t *Transaction) ReadPoolLock(allocID, blobberID string,
+	duration int64, lock, fee string) (err error) {
+	lv, err := parseCoinStr(lock)
+	if err != nil {
+		return err
+	}
+
+	fv, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type lockRequest struct {
+		Duration     time.Duration `json:"duration"`
+		AllocationID string        `json:"allocation_id"`
+		BlobberID    string        `json:"blobber_id,omitempty"`
+	}
+
+	var lr lockRequest
+	lr.Duration = time.Duration(duration)
+	lr.AllocationID = allocID
+	lr.BlobberID = blobberID
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_READ_POOL_LOCK, &lr, lv)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(fv)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// ReadPoolUnlock for current user and given pool.
+func (t *Transaction) ReadPoolUnlock(poolID string, fee string) (err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type unlockRequest struct {
+		PoolID string `json:"pool_id"`
+	}
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_READ_POOL_UNLOCK, &unlockRequest{
+			PoolID: poolID,
+		}, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// StakePoolLock used to lock tokens in a stake pool of a blobber.
+func (t *Transaction) StakePoolLock(blobberID string, lock, fee string) (
+	err error) {
+	lv, err := parseCoinStr(lock)
+	if err != nil {
+		return err
+	}
+
+	fv, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type stakePoolRequest struct {
+		BlobberID string `json:"blobber_id"`
+	}
+
+	var spr stakePoolRequest
+	spr.BlobberID = blobberID
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_STAKE_POOL_LOCK, &spr, lv)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(fv)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// StakePoolUnlock by blobberID and poolID.
+func (t *Transaction) StakePoolUnlock(blobberID, poolID string,
+	fee string) (err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type stakePoolRequest struct {
+		BlobberID string `json:"blobber_id"`
+		PoolID    string `json:"pool_id"`
+	}
+
+	var spr stakePoolRequest
+	spr.BlobberID = blobberID
+	spr.PoolID = poolID
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress, transaction.STORAGESC_STAKE_POOL_UNLOCK, &spr, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// UpdateBlobberSettings update settings of a blobber.
+func (t *Transaction) UpdateBlobberSettings(b *Blobber, fee string) (err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_UPDATE_BLOBBER_SETTINGS, b, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// UpdateAllocation transaction.
+func (t *Transaction) UpdateAllocation(allocID string, sizeDiff int64,
+	expirationDiff int64, lock, fee string) (err error) {
+	lv, err := parseCoinStr(lock)
+	if err != nil {
+		return err
+	}
+
+	fv, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type updateAllocationRequest struct {
+		ID         string `json:"id"`              // allocation id
+		Size       int64  `json:"size"`            // difference
+		Expiration int64  `json:"expiration_date"` // difference
+	}
+
+	var uar updateAllocationRequest
+	uar.ID = allocID
+	uar.Size = sizeDiff
+	uar.Expiration = expirationDiff
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_UPDATE_ALLOCATION, &uar, lv)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(fv)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// WritePoolLock locks tokens for current user and given allocation, using given
+// duration. If blobberID is not empty, then tokens will be locked for given
+// allocation->blobber only.
+func (t *Transaction) WritePoolLock(allocID, blobberID string, duration int64,
+	lock, fee string) (err error) {
+	lv, err := parseCoinStr(lock)
+	if err != nil {
+		return err
+	}
+
+	fv, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type lockRequest struct {
+		Duration     time.Duration `json:"duration"`
+		AllocationID string        `json:"allocation_id"`
+		BlobberID    string        `json:"blobber_id,omitempty"`
+	}
+
+	var lr lockRequest
+	lr.Duration = time.Duration(duration)
+	lr.AllocationID = allocID
+	lr.BlobberID = blobberID
+
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_WRITE_POOL_LOCK, &lr, lv)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(fv)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// WritePoolUnlock for current user and given pool.
+func (t *Transaction) WritePoolUnlock(poolID string, fee string) (
+	err error) {
+	v, err := parseCoinStr(fee)
+	if err != nil {
+		return err
+	}
+
+	type unlockRequest struct {
+		PoolID string `json:"pool_id"`
+	}
+	err = t.createSmartContractTxn(StorageSmartContractAddress,
+		transaction.STORAGESC_WRITE_POOL_UNLOCK, &unlockRequest{
+			PoolID: poolID,
+		}, 0)
+	if err != nil {
+		Logger.Error(err)
+		return
+	}
+	t.SetTransactionFee(v)
+	go func() { t.setNonceAndSubmit() }()
+	return
+}
+
+// ConvertToValue converts ZCN tokens to value
+func ConvertToValue(token float64) string {
+	return strconv.FormatUint(uint64(token*float64(TOKEN_UNIT)), 10)
+}
